% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBball.R
\name{runBball}
\alias{runBball}
\title{Basketball Shiny App}
\usage{
runBball()
}
\description{
This function launches a Shiny app called "basketball_app" included in the
package "mybasketball.kn". The app provides a plot and analysis of
college basketball data.
}
