% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home_vs_away_points.R
\name{home_vs_away_points}
\alias{home_vs_away_points}
\title{Create a ggplot box plot comparing the scores of home and away teams or a selected team's wins and losses.}
\usage{
home_vs_away_points(tibble, team = NULL)
}
\arguments{
\item{tibble}{A tibble containing game data with columns Score1 and Score2 representing
the scores of the home and away teams, respectively.}

\item{team}{(Optional) A character string specifying the team name for which the box plot should be created.
Defaults to NULL.}
}
\value{
A ggplot box plot comparing the scores of home and away teams or a selected team's wins and losses.
}
\description{
This function takes a tibble containing game data and creates a box plot
comparing the scores of home and away teams when no team is specified,
or creates a box plot showing the distribution of scores for a selected team's wins and losses.
}
\examples{
# Load necessary packages
library(tidyverse)

# Create sample game data
game_data <- tibble(
  Score1 = c(2, 3, 1, 2),
  Score2 = c(1, 2, 2, 1)
)

# Create the box plot of home vs. away scores
home_vs_away_points(game_data)

# Create the box plot of a selected team's wins and losses
home_vs_away_points(game_data, "Akron")

}
